globalvariablesの追加

+++ DirectXFunc.cpp
@@ -250,8 +250,8 @@ void DirectXFunc::PreDraw()
 	//ビューポート
 	D3D12_VIEWPORT viewport{};
 	//クライアント領域のサイズと一緒にして画面全体に表示
-	viewport.Width = WinApp::kClientWidth;
-	viewport.Height = WinApp::kClientHeight;
+	viewport.Width = (FLOAT)WinApp::kClientWidth;
+	viewport.Height = (FLOAT)WinApp::kClientHeight;
 	viewport.TopLeftX = 0;
 	viewport.TopLeftY = 0;
 	viewport.MinDepth = 0.0f;
